Lab 1

+++ src/main/java/tech/reliab/course/zimskovma/bank/service/BankService.java
@@ -1,7 +1,5 @@
 package tech.reliab.course.zimskovma.bank.service;
 
-import java.math.BigDecimal;
-
 import tech.reliab.course.zimskovma.bank.entity.Bank;
 
 
@@ -10,5 +8,5 @@ public interface BankService {
     Bank create(Bank bank);
 
     // Расчет процентной ставки банка
-    BigDecimal calculateInterestRate(Bank bank);
+    double calculateInterestRate(Bank bank);
 }
