+++ src/main/java/tech/reliab/course/zimskovma/bank/service/PaymentAccountService.java
@@ -7,7 +7,7 @@ import tech.reliab.course.zimskovma.bank.entity.PaymentAccount;
 public interface PaymentAccountService {
     PaymentAccount create(PaymentAccount paymentAccount);
 
-    boolean depositMoney(PaymentAccount paymentAccount, BigDecimal amount);
+    boolean depositMoney(PaymentAccount paymentAccount, double amount);
 
-    boolean withdrawMoney(PaymentAccount paymentAccount, BigDecimal amount);
+    boolean withdrawMoney(PaymentAccount paymentAccount, double amount);
 }
